+++ dashboard/src/pages/signin/email.tsx
@@ -8,9 +8,9 @@ import { Text } from '@/components/ui/v2/Text';
 import { getToastStyleProps } from '@/utils/constants/settings';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { styled } from '@mui/material';
-import { useSignInEmailPassword } from '@nhost/nextjs';
-import type { ReactElement } from 'react';
-import { useEffect } from 'react';
+import { useNhostClient, useSignInEmailPassword } from '@nhost/nextjs';
+import { useRouter } from 'next/router';
+import { useEffect, type ReactElement } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
 import { toast } from 'react-hot-toast';
 import * as Yup from 'yup';
@@ -30,6 +30,8 @@ const StyledInput = styled(Input)({
 });
 
 export default function EmailSignUpPage() {
+  const router = useRouter();
+  const nhost = useNhostClient();
   const { signInEmailPassword, error } = useSignInEmailPassword();
 
   const form = useForm<EmailSignUpFormValues>({
@@ -56,7 +58,15 @@ export default function EmailSignUpPage() {
 
   async function handleSubmit({ email, password }: EmailSignUpFormValues) {
     try {
-      await signInEmailPassword(email, password);
+      const { needsEmailVerification } = await signInEmailPassword(
+        email,
+        password,
+      );
+
+      if (needsEmailVerification) {
+        await nhost.auth.sendVerificationEmail({ email: email as string });
+        router.push(`/email/verify?email=${email}`);
+      }
     } catch {
       toast.error(
         'An error occurred while signing in. Please try again.',
@@ -75,7 +85,7 @@ export default function EmailSignUpPage() {
         Sign In
       </Text>
 
-      <Box className="grid grid-flow-row gap-4 rounded-md border bg-transparent p-6 lg:p-12">
+      <Box className="grid grid-flow-row gap-4 p-6 bg-transparent border rounded-md lg:p-12">
         <FormProvider {...form}>
           <Form
             onSubmit={handleSubmit}
@@ -115,7 +125,7 @@ export default function EmailSignUpPage() {
             <NavLink
               href="/reset-password"
               color="white"
-              className="justify-self-start font-semibold"
+              className="font-semibold justify-self-start"
             >
               Forgot password?
             </NavLink>
@@ -140,7 +150,7 @@ export default function EmailSignUpPage() {
         </FormProvider>
       </Box>
 
-      <Text color="secondary" className="text-center text-base lg:text-lg">
+      <Text color="secondary" className="text-base text-center lg:text-lg">
         Don&apos;t have an account?{' '}
         <NavLink href="/signup" color="white">
           Sign Up