+++ examples/quickstarts/nextjs-server-components/.eslintrc.js
@@ -1,6 +1,7 @@
 module.exports = {
   extends: ['../../../config/.eslintrc.js', 'plugin:@next/next/recommended'],
   rules: {
-    'react/react-in-jsx-scope': 'off'
+    'react/react-in-jsx-scope': 'off',
+    '@next/next/no-server-import-in-page': 'off'
   }
 }