+++ examples/quickstarts/nextjs-server-components/src/middleware.ts
@@ -1,6 +1,4 @@
 import { manageAuthSession } from '@utils/nhost'
-
-// eslint-disable-next-line @next/next/no-server-import-in-page
 import { NextRequest, NextResponse } from 'next/server'
 
 export async function middleware(request: NextRequest) {