+++ examples/codegen-react-apollo/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @nhost-examples/codegen-react-apollo
 
+## 0.4.1
+
+### Patch Changes
+
+- @nhost/react@3.4.1
+- @nhost/react-apollo@11.0.1
+
+## 0.4.0
+
+### Minor Changes
+
+- 768ca17: chore: update dependencies
+
+### Patch Changes
+
+- Updated dependencies [768ca17]
+  - @nhost/react-apollo@11.0.0
+  - @nhost/react@3.4.0
+
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies [311374e]
+  - @nhost/react@3.3.2
+  - @nhost/react-apollo@10.0.2
+
 ## 0.3.0
 
 ### Minor Changes