+++ examples/codegen-react-query/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @nhost-examples/codegen-react-query
 
+## 0.4.1
+
+### Patch Changes
+
+- @nhost/react@3.4.1
+
+## 0.4.0
+
+### Minor Changes
+
+- 768ca17: chore: update dependencies
+
+### Patch Changes
+
+- Updated dependencies [768ca17]
+  - @nhost/react@3.4.0
+
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies [311374e]
+  - @nhost/react@3.3.2
+
 ## 0.3.0
 
 ### Minor Changes