+++ examples/nextjs/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @nhost-examples/nextjs
 
+## 0.3.1
+
+### Patch Changes
+
+- @nhost/react@3.4.1
+- @nhost/react-apollo@11.0.1
+- @nhost/nextjs@2.1.10
+
+## 0.3.0
+
+### Minor Changes
+
+- 768ca17: chore: update dependencies
+
+### Patch Changes
+
+- Updated dependencies [768ca17]
+  - @nhost/react-apollo@11.0.0
+  - @nhost/react@3.4.0
+  - @nhost/nextjs@2.1.9
+
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies [311374e]
+  - @nhost/react@3.3.2
+  - @nhost/react-apollo@10.0.2
+  - @nhost/nextjs@2.1.8
+
 ## 0.2.1
 
 ### Patch Changes