+++ examples/node-storage/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @nhost-examples/node-storage
 
+## 0.2.1
+
+### Patch Changes
+
+- @nhost/nhost-js@3.0.11
+
+## 0.2.0
+
+### Minor Changes
+
+- 768ca17: chore: update dependencies
+
+### Patch Changes
+
+- @nhost/nhost-js@3.0.10
+
+## 0.1.2
+
+### Patch Changes
+
+- @nhost/nhost-js@3.0.10
+
 ## 0.1.1
 
 ### Patch Changes