+++ examples/react-gqty/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @nhost-examples/react-gqty
 
+## 1.2.1
+
+### Patch Changes
+
+- @nhost/react@3.4.1
+
+## 1.2.0
+
+### Minor Changes
+
+- 768ca17: chore: update dependencies
+
+### Patch Changes
+
+- Updated dependencies [768ca17]
+  - @nhost/react@3.4.0
+
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies [311374e]
+  - @nhost/react@3.3.2
+
 ## 1.1.1
 
 ### Patch Changes