+++ integrations/apollo/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @nhost/apollo
 
+## 6.2.1
+
+### Patch Changes
+
+- @nhost/nhost-js@3.0.11
+
+## 6.2.0
+
+### Minor Changes
+
+- 768ca17: chore: update dependencies
+
+### Patch Changes
+
+- @nhost/nhost-js@3.0.10
+
+## 6.1.2
+
+### Patch Changes
+
+- @nhost/nhost-js@3.0.10
+
 ## 6.1.1
 
 ### Patch Changes