+++ integrations/react-apollo/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @nhost/react-apollo
 
+## 11.0.1
+
+### Patch Changes
+
+- @nhost/apollo@6.2.1
+- @nhost/react@3.4.1
+
+## 11.0.0
+
+### Minor Changes
+
+- 768ca17: chore: update dependencies
+
+### Patch Changes
+
+- Updated dependencies [768ca17]
+  - @nhost/apollo@6.2.0
+  - @nhost/react@3.4.0
+
+## 10.0.2
+
+### Patch Changes
+
+- Updated dependencies [311374e]
+  - @nhost/react@3.3.2
+  - @nhost/apollo@6.1.2
+
 ## 10.0.1
 
 ### Patch Changes