+++ integrations/react-urql/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @nhost/react-urql
 
+## 8.0.1
+
+### Patch Changes
+
+- @nhost/react@3.4.1
+
+## 8.0.0
+
+### Minor Changes
+
+- 768ca17: chore: update dependencies
+
+### Patch Changes
+
+- Updated dependencies [768ca17]
+  - @nhost/react@3.4.0
+
+## 7.0.2
+
+### Patch Changes
+
+- Updated dependencies [311374e]
+  - @nhost/react@3.3.2
+
 ## 7.0.1
 
 ### Patch Changes