+++ packages/docgen/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nhost/docgen
 
+## 0.3.0
+
+### Minor Changes
+
+- 768ca17: chore: update dependencies
+
 ## 0.2.0
 
 ### Minor Changes