+++ packages/nextjs/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @nhost/nextjs
 
+## 2.1.10
+
+### Patch Changes
+
+- @nhost/react@3.4.1
+
+## 2.1.9
+
+### Patch Changes
+
+- Updated dependencies [768ca17]
+  - @nhost/react@3.4.0
+
+## 2.1.8
+
+### Patch Changes
+
+- Updated dependencies [311374e]
+  - @nhost/react@3.3.2
+
 ## 2.1.7
 
 ### Patch Changes