+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @nhost/vue
 
+## 2.5.1
+
+### Patch Changes
+
+- @nhost/nhost-js@3.0.11
+
+## 2.5.0
+
+### Minor Changes
+
+- 768ca17: chore: update dependencies
+
+### Patch Changes
+
+- @nhost/nhost-js@3.0.10
+
+## 2.4.0
+
+### Minor Changes
+
+- 311374e: feat: add `connect` param to `useProviderLink` hook
+
+### Patch Changes
+
+- @nhost/nhost-js@3.0.10
+
 ## 2.3.1
 
 ### Patch Changes